#48 Dodanie pola "Numer Transakcji", wyśrodkowanie pól tekstowych.
