+++ SmartShop/SmartShopWpf/Hints.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Windows.Data;
+using System.Windows;
+
+
+namespace SmartShopWpf
+{
+   public class Hints : IMultiValueConverter
+    { 
+     public object Convert(object[] values, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+    {
+        // Always test MultiValueConverter inputs for non-null
+        // (to avoid crash bugs for views in the designer)
+        if (values[0] is bool && values[1] is bool)
+        {
+            bool hasText = !(bool)values[0];
+            bool hasFocus = (bool)values[1];
+
+            if (hasFocus || hasText)
+                return Visibility.Collapsed;
+        }
+
+        return Visibility.Visible;
+    }
+
+
+    public object[] ConvertBack(object value, Type[] targetTypes, object parameter, System.Globalization.CultureInfo culture)
+    {
+        throw new NotImplementedException();
+    }
+
+
+}
+}
