--- SmartShop/SmartShopWpf/LoginWindow.cs
@@ -1,34 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-
-namespace SmartShopWpf
-{
-    /// <summary>
-    /// Interaction logic for OknoLogowania.xaml
-    /// </summary>
-    public partial class OknoLogowania : Window
-    {
-        public OknoLogowania()
-        {
-            InitializeComponent();
-        }
-
-        private void btnZaloguj_Click(object sender, RoutedEventArgs e)
-        {
-            MainWindow mW = new MainWindow();
-            mW.Show();
-            this.Close();
-        }
-    }
-}
