+++ SmartShop/SmartShopWpf/LoginWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="SmartShopWpf.OknoLogowania"
+﻿<Window x:Class="SmartShopWpf.LoginWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -6,7 +6,13 @@
         xmlns:local="clr-namespace:SmartShopWpf"
         mc:Ignorable="d"
         ResizeMode="NoResize"
-        Title="OknoLogowania" Height="729" Width="1131">
+        Title="Okno Logowania" Height="729" Width="1131">
+
+    <Window.Resources>
+        <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
+        <local:LoginWindow x:Key="Hints" />
+    </Window.Resources>
+
 
     <Grid Height="698" VerticalAlignment="Bottom" HorizontalAlignment="Left" Width="1123" Background="{StaticResource Frame}">
 
@@ -15,8 +21,11 @@
         </Grid.ColumnDefinitions>
         <Label  x:Name="lblLogin"      HorizontalAlignment="Left" VerticalAlignment="Top" Height="73"  Width="999"  Margin="60,37,0,0" Background="{StaticResource Orange}" BorderThickness="1,1,1,1" BorderBrush="{StaticResource White}" Content="Logowanie" Style="{StaticResource LblFont36}" HorizontalContentAlignment="Center" VerticalContentAlignment="Center"/>
 
+
+      
         <TextBox x:Name="txtLogin"     HorizontalAlignment="Left" VerticalAlignment="Top" Height="55" Width="301" Margin="413,310,0,0" Background="{StaticResource Orange}" BorderThickness="1,1,1,1" BorderBrush="{StaticResource White}" Style="{StaticResource TxtFont26}" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" />
-        <TextBox x:Name="txtPassword"  HorizontalAlignment="Left" VerticalAlignment="Top" Height="55" Width="301" Margin="413,364,0,0" Background="{StaticResource Orange}" BorderThickness="1,1,1,1" BorderBrush="{StaticResource White}" Style="{StaticResource TxtFont26}" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" />
+     
+        <TextBox x:Name="txtPassword"  HorizontalAlignment="Left" VerticalAlignment="Top" Height="55" Width="301" Margin="413,364,0,0" Background="{StaticResource Orange}" BorderThickness="1,1,1,1" BorderBrush="{StaticResource White}" Style="{StaticResource TxtFont26}" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" Text="Hasło" />
 
 
         <Button x:Name="btnLogin"      HorizontalAlignment="Left" VerticalAlignment="Top" Height="55" Width="301" Margin="413,418,0,0" Background="{StaticResource Red}"    BorderThickness="1,1,1,1" BorderBrush="{StaticResource White}" Content="Zaloguj"         Style="{StaticResource BtnFontRed30}" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" Click="btnLogin_Click"/>
