+++ SmartShop/SmartShopWpf/SmartShopWpf.csproj
@@ -69,14 +69,14 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Page Include="OknoLogowania.xaml">
+    <Page Include="LoginWindow.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="OknoLogowania.xaml.cs">
-      <DependentUpon>OknoLogowania.xaml</DependentUpon>
+    <Compile Include="LoginWindow.xaml.cs">
+      <DependentUpon>LoginWindow.xaml</DependentUpon>
     </Compile>
     <Compile Include="Models\Cashbox.cs" />
     <Compile Include="Models\Cashier.cs" />
